Move exceptions to own files
