+++ common/Exceptions/ForbidenException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Valued\Magento2\Exceptions;
+
+class ForbidenException extends ProductReviewSyncException {
+    public function getHttpResponseCode(): int {
+        return 403;
+    }
+}
