+++ common/Exceptions/MethodNotAllowed.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Valued\Magento2\Exceptions;
+
+class MethodNotAllowed extends ProductReviewSyncException {
+    public function getHttpResponseCode(): int {
+        return 405;
+    }
+}
