+++ common/Exceptions/ProductReviewSyncException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Valued\Magento2\Exceptions;
+
+class ProductReviewSyncException extends \Exception {
+
+}
