+++ common/Exceptions/UnauthorizedException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Valued\Magento2\Exceptions;
+
+class UnauthorizedException extends ProductReviewSyncException {
+    public function getHttpResponseCode(): int {
+        return 401;
+    }
+}
