+++ common/Exceptions/UnconfiguredAppException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace Valued\Magento2\Exceptions;
+
+class UnconfiguredAppException extends ProductReviewSyncException {
+    public function getHttpResponseCode(): int {
+        return 501;
+    }
+}
